fix: widen github_issue_id to BIGINT — integer out of range for large IDs
GitHub issue IDs have exceeded 2^31 (~2.1B). IDs like 4249778080 cause
psycopg2 NumericValueOutOfRange on INSERT. Fix:
  - New migration 0002 ALTERs tasks.github_issue_id from INTEGER to BIGINT
  - Initial schema updated to BIGINT for fresh deployments
  - Task model updated to BigInteger to match

+++ backend/migrations/versions/0001_initial_schema.py
@@ -106,7 +106,7 @@ def upgrade() -> None:
             id                    UUID                     NOT NULL PRIMARY KEY,
             project_id            UUID                     NOT NULL
                                       REFERENCES projects (id) ON DELETE CASCADE,
-            github_issue_id       INTEGER UNIQUE,
+            github_issue_id       BIGINT UNIQUE,
             github_issue_number   INTEGER,
             title                 VARCHAR(500)             NOT NULL,
             description_original  TEXT,
